feat: 커뮤니티 API 응답 추가 (#68)

+++ server/pulsar/src/main/java/com/hammer/pulsar/service/InteractionServiceImpl.java
@@ -7,10 +7,12 @@ import com.hammer.pulsar.dto.interaction.Comment;
 import com.hammer.pulsar.dto.interaction.CommentWriteRequest;
 import com.hammer.pulsar.dto.interaction.Like;
 import com.hammer.pulsar.dto.interaction.LikeRequest;
+import com.hammer.pulsar.exception.UnauthorizedException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 // 실제 로직이 구현된 InteractionService 인터페이스의 구현체 클래스
 @Service
@@ -39,7 +41,7 @@ public class InteractionServiceImpl implements InteractionService {
         // 추천 여부 불러오기
         result.setLiked(likeDao.selectLikeByMemberId(request));
         // 총 추천수 불러오기
-        result.setLikes(likeDao.selectLikes(request.getArticleId()));
+        result.setLikes(articleDao.countLikes(request.getArticleId()));
 
         return result;
     }
@@ -60,7 +62,7 @@ public class InteractionServiceImpl implements InteractionService {
 
         // 추천 상태 토글하고 결과 반환
         result.setLiked(toggleStatus(request, liked));
-        result.setLikes(likeDao.selectLikes(request.getArticleId()));
+        result.setLikes(articleDao.countLikes(request.getArticleId()));
 
         return result;
     }
@@ -118,7 +120,12 @@ public class InteractionServiceImpl implements InteractionService {
      * @return 삭제 후의 댓글 목록을 반환한다.
      */
     @Override
-    public List<Comment> removeComment(int articleId, int commentId) {
+    public List<Comment> removeComment(int articleId, int commentId, int memberId) {
+        Comment comment = commentDao.selectCommentByCommentId(commentId);
+
+        if(comment == null) throw new NoSuchElementException("존재하지 않는 댓글입니다.");
+        if(comment.getWriterInfo().getWriterNo() != memberId) throw new UnauthorizedException("권한이 없습니다.");
+
         commentDao.deleteComment(commentId);
 
         return commentDao.selectCommentsByArticleId(articleId);
